+++ src/components/app-header/app-header.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+export default class AppHeader extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.state = {}        
+    }
+
+    render() {
+        return(
+            <h1>HEADER</h1>
+        )
+    }
+
+}
