+++ src/components/app/app.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import AppHeader from '../app-header/app-header';
+
+function App() {
+  return (
+    <div className="App">
+      <AppHeader />
+    </div>
+  );
+}
+
+export default App;
